feat: adds reading of data from json files
Implement tests to verify that the json file is successfully loaded and read.

Scenarios verified with tests are:

[SUCCESS]
- Search for the data.json file and return its content.

[ERROR]
- Return an empty array and log an error if the search fails (connection problem).
- Return an empty array and log an error if the HTTP response is different from ok (200).

The npm test command has been updated to allow the use of ESM syntax with Jest.

To follow good development practices, the application has gained two modules:

[app]
The application entry point, which contains the system initialization logic.

[dataHandler]
Responsible for I/O operations on json files.